Version 20.63:    Apr 16, 2021  Optimizations in database selects.

+++ swad_exam_result.c
@@ -427,28 +427,29 @@ void ExaRes_ShowAllResultsInExa (void)
 static void ExaRes_ListAllResultsInExa (struct Exa_Exams *Exams,long ExaCod)
   {
    MYSQL_RES *mysql_res;
-   unsigned long NumUsrs;
-   unsigned long NumUsr;
+   unsigned NumUsrs;
+   unsigned NumUsr;
 
    /***** Table head *****/
    ExaRes_ShowHeaderResults (Usr_OTHER);
 
    /***** Get all users who have answered any session question in this exam *****/
-   NumUsrs = DB_QuerySELECT (&mysql_res,"can not get users in exam",
-			     "SELECT users.UsrCod"	// row[0]
-			      " FROM (SELECT DISTINCT exa_prints.UsrCod AS UsrCod"
-			              " FROM exa_prints,exa_sessions,exa_exams"
-			             " WHERE exa_sessions.ExaCod=%ld"
-			               " AND exa_sessions.SesCod=exa_prints.SesCod"
-			               " AND exa_sessions.ExaCod=exa_exams.ExaCod"
-			               " AND exa_exams.CrsCod=%ld) AS users,"		// Extra check
-			            "usr_data"
-			     " WHERE users.UsrCod=usr_data.UsrCod"
-			     " ORDER BY usr_data.Surname1,"
-			               "usr_data.Surname2,"
-			               "usr_data.FirstName",
-			     ExaCod,
-			     Gbl.Hierarchy.Crs.CrsCod);
+   NumUsrs = (unsigned)
+   DB_QuerySELECT (&mysql_res,"can not get users in exam",
+		   "SELECT users.UsrCod"	// row[0]
+		    " FROM (SELECT DISTINCT exa_prints.UsrCod AS UsrCod"
+			    " FROM exa_prints,exa_sessions,exa_exams"
+			   " WHERE exa_sessions.ExaCod=%ld"
+			     " AND exa_sessions.SesCod=exa_prints.SesCod"
+			     " AND exa_sessions.ExaCod=exa_exams.ExaCod"
+			     " AND exa_exams.CrsCod=%ld) AS users,"		// Extra check
+			  "usr_data"
+		   " WHERE users.UsrCod=usr_data.UsrCod"
+		   " ORDER BY usr_data.Surname1,"
+			     "usr_data.Surname2,"
+			     "usr_data.FirstName",
+		   ExaCod,
+		   Gbl.Hierarchy.Crs.CrsCod);
 
    /***** List sessions results for each user *****/
    for (NumUsr = 0;
@@ -516,28 +517,29 @@ void ExaRes_ShowAllResultsInSes (void)
 static void ExaRes_ListAllResultsInSes (struct Exa_Exams *Exams,long SesCod)
   {
    MYSQL_RES *mysql_res;
-   unsigned long NumUsrs;
-   unsigned long NumUsr;
+   unsigned NumUsrs;
+   unsigned NumUsr;
 
    /***** Table head *****/
    ExaRes_ShowHeaderResults (Usr_OTHER);
 
    /***** Get all users who have answered any session question in this exam *****/
-   NumUsrs = DB_QuerySELECT (&mysql_res,"can not get users in session",
-			     "SELECT users.UsrCod"	// row[0]
-			      " FROM (SELECT exa_prints.UsrCod AS UsrCod"	// row[0]
-				      " FROM exa_prints,exa_sessions,exa_exams"
-				     " WHERE exa_prints.SesCod=%ld"
-				       " AND exa_prints.SesCod=exa_sessions.SesCod"
-				       " AND exa_sessions.ExaCod=exa_exams.ExaCod"
-				       " AND exa_exams.CrsCod=%ld) AS users,"	// Extra check
-			            "usr_data"
-			     " WHERE users.UsrCod=usr_data.UsrCod"
-			     " ORDER BY usr_data.Surname1,"
-			               "usr_data.Surname2,"
-			               "usr_data.FirstName",
-			     SesCod,
-			     Gbl.Hierarchy.Crs.CrsCod);
+   NumUsrs = (unsigned)
+   DB_QuerySELECT (&mysql_res,"can not get users in session",
+		   "SELECT users.UsrCod"	// row[0]
+		    " FROM (SELECT exa_prints.UsrCod AS UsrCod"	// row[0]
+			    " FROM exa_prints,exa_sessions,exa_exams"
+			   " WHERE exa_prints.SesCod=%ld"
+			     " AND exa_prints.SesCod=exa_sessions.SesCod"
+			     " AND exa_sessions.ExaCod=exa_exams.ExaCod"
+			     " AND exa_exams.CrsCod=%ld) AS users,"	// Extra check
+			  "usr_data"
+		   " WHERE users.UsrCod=usr_data.UsrCod"
+		   " ORDER BY usr_data.Surname1,"
+			     "usr_data.Surname2,"
+			     "usr_data.FirstName",
+		   SesCod,
+		   Gbl.Hierarchy.Crs.CrsCod);
 
    /***** List sessions results for each user *****/
    for (NumUsr = 0;
@@ -905,25 +907,25 @@ static void ExaRes_ShowResults (struct Exa_Exams *Exams,
    /***** Make database query *****/
    // Do not filter by groups, because a student who has changed groups
    // must be able to access exams taken in other groups
-   NumResults =
-   (unsigned) DB_QuerySELECT (&mysql_res,"can not get sessions results",
-			      "SELECT exa_prints.PrnCod"			// row[0]
-			       " FROM exa_prints,exa_sessions,exa_exams"
-			      " WHERE exa_prints.UsrCod=%ld"
-			         "%s"	// Session subquery
-			        " AND exa_prints.SesCod=exa_sessions.SesCod"
-                                 "%s"	// Hidden sessions subquery
-			         "%s"	// Exams subquery
-			        " AND exa_sessions.ExaCod=exa_exams.ExaCod"
-                                 "%s"	// Hidden exams subquery
-			        " AND exa_exams.CrsCod=%ld"			// Extra check
-			      " ORDER BY exa_sessions.Title",
-			      UsrDat->UsrCod,
-			      SesSubQuery,
-			      HidSesSubQuery,
-			      ExaSubQuery,
-			      HidExaSubQuery,
-			      Gbl.Hierarchy.Crs.CrsCod);
+   NumResults = (unsigned)
+   DB_QuerySELECT (&mysql_res,"can not get sessions results",
+		   "SELECT exa_prints.PrnCod"			// row[0]
+		    " FROM exa_prints,exa_sessions,exa_exams"
+		   " WHERE exa_prints.UsrCod=%ld"
+		      "%s"	// Session subquery
+		     " AND exa_prints.SesCod=exa_sessions.SesCod"
+		      "%s"	// Hidden sessions subquery
+		      "%s"	// Exams subquery
+		     " AND exa_sessions.ExaCod=exa_exams.ExaCod"
+		      "%s"	// Hidden exams subquery
+		     " AND exa_exams.CrsCod=%ld"			// Extra check
+		   " ORDER BY exa_sessions.Title",
+		   UsrDat->UsrCod,
+		   SesSubQuery,
+		   HidSesSubQuery,
+		   ExaSubQuery,
+		   HidExaSubQuery,
+		   Gbl.Hierarchy.Crs.CrsCod);
    free (HidExaSubQuery);
    free (ExaSubQuery);
    free (HidSesSubQuery);
